saltapi/service/user.py
from dataclasses import dataclass
from enum import Enum
from typing import NamedTuple, Optional, List


class Role(str, Enum):
    SALT_ASTRONOMER = "SALT Astronomer"
    ADMINISTRATOR = "Administrator"
    TAC_MEMBER = "TAC Member"
    TAC_CHAIR = "TAC Chair"
    BOARD_MEMBER = "Board Member"


@dataclass()
class User:
    id: int
    username: str
    given_name: str
    family_name: str
    email: str
    password_hash: str
    roles: List[Role]


class UserUpdate(NamedTuple):
    username: Optional[str]
    # Not implemented yet
    # given_name: Optional[str]
    # family_name: Optional[str]
    # email: Optional[str]
    password: Optional[str]
